Broke the citizen list and detail classes into two separate files, as suggested.
